fix: don't fork(), exec() and wait(), but just exec()
This change causes the perl process to not hang around during the
execution of OpenResty, which is particularly notable when using
OpenResty to start something long-running in the foreground.

The code for forwarding the signals to the child process has been
removed in the process.  Signals will be handled by the OpenResty
process as normal instead of through the intermediate process that
forwards them.